Introducing UseEqual feature
If set to true (via DefaultContextConfig.UseEqual or using
T.UseEqual), use the Equal method on got (if it exists) or on any of
its component to compare got and expected values.

The signature of the Equal method should be:
  (A) Equal(B) bool
with B assignable to A.

See time.Time as an example of accepted Equal() method.

Signed-off-by: Maxime Soulé <btik-git@scoubidou.com>