github.com/maxatome/go-testdeep → github.com/maxatome/go-testdeep/td
github.com/maxatome/go-testdeep still usable so the compatibility is
preserved.

The only breaking change is the disappearance of
testdeep.DefaultContextConfig. It should be replaced by
td.DefaultContextConfig and so the go-testdeep import line should be
changed (at least "github.com/maxatome/go-testdeep/td" should be
added).

To summarize: migration to github.com/maxatome/go-testdeep/to is not
needed, only recommended, except if testdeep.DefaultContextConfig is
used.

To migrate:

    import "github.com/maxatome/go-testdeep"
    …
    testdeep.DefaultContextConfig = testdeep.ContextConfig{…}
    testdeep.Cmp(…)

or with "td" package alias:

    import td "github.com/maxatome/go-testdeep"
    …
    td.DefaultContextConfig = td.ContextConfig{…}
    td.Cmp(…)

simply do (see import line changes):

    import "github.com/maxatome/go-testdeep/td"
    …
    td.DefaultContextConfig = td.ContextConfig{…}
    td.Cmp(…)

Signed-off-by: Maxime Soulé <btik-git@scoubidou.com>