feat: don't panic on users errors if unsafe available and not -race
but call (*testing.T).Fatal().

To achieve this, we play with unsafe.Pointer to *testing.T casting,
but data race detection panics:

  fatal error: checkptr: pointer arithmetic result points to invalid allocation

To disable this panic, one can use the following compilation flag:

  -gcflags=all=-d=checkptr=0

but we don't want go-testdeep users to be forced to set this flag in
their workflow.

Until we find a better solution, go-testdeep continues to panic on
users errors when data race detection (-race) is enabled.

Based on Dave Cheney idea at
https://github.com/pkg/expect/blob/1fe4c9394a8a0ef4038dc997ea676f44784b7bae/t.go

Signed-off-by: Maxime Soulé <btik-git@scoubidou.com>