feat: add (*td.T).Parallel() method
Add a Parallel() method to `*td.T` that calls `Parallel()` on the
underlying `testing.TB` if it supports the method. This will have no
effect on `*testing.B` or custom wrappers that do not support
`Parallel()`.

Per #149, this implements this as a method rather than a top-level
`td.Parallel(t)` function because this is closer to what existing usage
of `*testing.T` looks like.

Resolves #149